move generational workqueue stream into available api
